Rename FemElasticity to ElasticityElement.
